requirements.txt
beautifulsoup4==4.11.1
disnake==2.5.1
python-dotenv==0.20.0
requests==2.28.1
